Refactor krnlmon code
