+++ switchlink/switchlink_neigh.h
@@ -21,7 +21,8 @@
 #include <stdbool.h>
 
 struct nlmsghdr;
-extern void neigh_create(switchlink_handle_t vrf_h,
+/*
+extern void switchlink_create_neigh(switchlink_handle_t vrf_h,
                          switchlink_ip_addr_t *ipaddr,
                          switchlink_mac_addr_t mac_addr,
                          switchlink_handle_t intf_h);
@@ -30,7 +31,7 @@ extern void switchlink_linux_mac_update(switchlink_mac_addr_t mac_addr,
                                         switchlink_handle_t bridge_h,
                                         switchlink_handle_t intf_h,
                                         bool create);
-
+*/
 void process_neigh_msg(struct nlmsghdr *nlmsg, int type);
 
 #endif /* __SWITCHLINK_NEIGH_H__ */
