+++ switchlink/switchlink_route.c
@@ -15,224 +15,9 @@
  * limitations under the License.
  */
 
-#include <stdint.h>
-#include <stdbool.h>
-#include <netlink/netlink.h>
-#include <netlink/msg.h>
-
 #include "config.h"
-#include "switchlink.h"
-#include "switchlink_link.h"
-#include "switchlink_neigh.h"
+//#include "switchlink_neigh.h"
 #include "switchlink_route.h"
-#include "switchlink_db.h"
-#include "switchlink_sai.h"
-
-/*
- * Routine Description:
- *    Delete ecmp and remove entry to the database
- *
- * Arguments:
- *    [in] ecmp_h - route interface handle
- *
- * Return Values:
- *    void
- */
-
-static void ecmp_delete(switchlink_handle_t ecmp_h) {
-  int32_t ref_count;
-  uint32_t index = 0;
-  uint8_t num_nhops = 0;
-  switchlink_db_status_t status;
-  switchlink_db_nexthop_info_t nexthop_info;
-  switchlink_handle_t nhops[SWITCHLINK_ECMP_NUM_MEMBERS_MAX] = {0};
-
-  status = switchlink_db_ecmp_ref_dec(ecmp_h, &ref_count);
-  krnlmon_assert(status == SWITCHLINK_DB_STATUS_SUCCESS);
-
-  if (ref_count == 0) {
-    switchlink_db_ecmp_info_t ecmp_info;
-    memset(&ecmp_info, 0, sizeof(switchlink_db_ecmp_info_t));
-    status = switchlink_db_ecmp_handle_get_info(ecmp_h, &ecmp_info);
-    krnlmon_assert(status == SWITCHLINK_DB_STATUS_SUCCESS);
-    num_nhops = ecmp_info.num_nhops;
-    for (index = 0; index < num_nhops; index++) {
-      nhops[index] = ecmp_info.nhops[index];
-    }
-    dzlog_info("Deleting ecmp handler 0x%lx", ecmp_h);
-    switchlink_ecmp_delete(&ecmp_info);
-    switchlink_db_ecmp_delete(ecmp_h);
-
-    for (index = 0; index < num_nhops; index++) {
-      memset(&nexthop_info, 0, sizeof(switchlink_db_nexthop_info_t));
-      status = switchlink_db_nexthop_handle_get_info(nhops[index],
-                                                     &nexthop_info);
-      if (status != SWITCHLINK_DB_STATUS_SUCCESS) {
-        dzlog_error("Cannot get nhop info for nhop handle 0x%lx", nhops[index]);
-        continue;
-      }
-
-      if (validate_nexthop_delete(nexthop_info.using_by,
-                                  SWITCHLINK_NHOP_FROM_ROUTE)) {
-        dzlog_debug("Deleting nhop 0x%lx, from ecmp_delete", nexthop_info.nhop_h);
-        switchlink_nexthop_delete(nexthop_info.nhop_h);
-        switchlink_db_nexthop_delete(&nexthop_info);
-      } else {
-          dzlog_debug("Removing Route learn from nhop");
-        nexthop_info.using_by &= ~SWITCHLINK_NHOP_FROM_ROUTE;
-        switchlink_db_nexthop_update_using_by(&nexthop_info);
-      }
-    }
-  }
-}
-
-/*
- * Routine Description:
- *    Create route and add entry to the database
- *
- * Arguments:
- *    [in] vrf_h - vrf handle
- *    [in] dst - IP address associated with route
- *    [in] gateway - gateway associated with route
- *    [in] ecmp_h - route interface handle
- *    [in] intf_h - ecmp handle
- *
- * Return Values:
- *    void
- */
-
-void route_create(switchlink_handle_t vrf_h,
-                  switchlink_ip_addr_t *dst,
-                  switchlink_ip_addr_t *gateway,
-                  switchlink_handle_t ecmp_h,
-                  switchlink_handle_t intf_h) {
-  if (!dst || (!gateway && !ecmp_h)) {
-    if (ecmp_h) {
-      ecmp_delete(ecmp_h);
-    }
-    return;
-  }
-
-  bool ecmp_valid = false;
-  switchlink_handle_t nhop_h = g_cpu_rx_nhop_h;
-  if (!ecmp_h) {
-    // Ignore NULL gateway address, dont create a NHOP for that
-    if (gateway->ip.v4addr.s_addr) {
-      switchlink_db_nexthop_info_t nexthop_info;
-      memset(&nexthop_info, 0, sizeof(switchlink_db_nexthop_info_t));
-      memcpy(&(nexthop_info.ip_addr), gateway, sizeof(switchlink_ip_addr_t));
-      nexthop_info.intf_h = intf_h;
-      nexthop_info.vrf_h = vrf_h;
-      switchlink_db_status_t status;
-      status = switchlink_db_nexthop_get_info(&nexthop_info);
-      if (status == SWITCHLINK_DB_STATUS_SUCCESS) {
-          dzlog_debug("Received nhop 0x%lx handler, update from"
-                   " route", nexthop_info.nhop_h);
-        nhop_h = nexthop_info.nhop_h;
-        nexthop_info.using_by |= SWITCHLINK_NHOP_FROM_ROUTE;
-        switchlink_db_nexthop_update_using_by(&nexthop_info);
-      } else {
-        if (!switchlink_nexthop_create(&nexthop_info)) {
-          dzlog_debug("Created nhop 0x%lx handle, update from "
-                   " route", nexthop_info.nhop_h);
-          nhop_h = nexthop_info.nhop_h;
-          nexthop_info.using_by |= SWITCHLINK_NHOP_FROM_ROUTE;
-          switchlink_db_nexthop_add(&nexthop_info);
-        }
-      }
-    }
-    ecmp_valid = false;
-  } else {
-    ecmp_valid = true;
-    nhop_h = ecmp_h;
-  }
-
-  // get the route from the db (if it already exists)
-  switchlink_db_route_info_t route_info;
-  memset(&route_info, 0, sizeof(switchlink_db_route_info_t));
-  route_info.vrf_h = vrf_h;
-  memcpy(&(route_info.ip_addr), dst, sizeof(switchlink_ip_addr_t));
-  switchlink_db_status_t status = switchlink_db_route_get_info(&route_info);
-  if (status == SWITCHLINK_DB_STATUS_SUCCESS) {
-    if ((route_info.ecmp == ecmp_valid) && (route_info.nhop_h == nhop_h)) {
-      // no change
-      return;
-    }
-    // nexthop has change, delete the current route
-    route_delete(vrf_h, dst);
-  }
-
-  memset(&route_info, 0, sizeof(switchlink_db_route_info_t));
-  route_info.vrf_h = vrf_h;
-  memcpy(&(route_info.ip_addr), dst, sizeof(switchlink_ip_addr_t));
-  route_info.ecmp = ecmp_valid;
-  route_info.nhop_h = nhop_h;
-  route_info.intf_h = intf_h;
-
-  // add the route
-  dzlog_info("Create route: 0x%x/%d", dst->ip.v4addr.s_addr,
-             dst->prefix_len);
-  if (switchlink_route_create(&route_info) == -1) {
-    if (route_info.ecmp) {
-      ecmp_delete(route_info.nhop_h);
-    }
-    return;
-  }
-
-  // add the route to the db
-  if (switchlink_db_route_add(&route_info) == SWITCHLINK_DB_STATUS_SUCCESS) {
-    if (route_info.ecmp) {
-      switchlink_db_ecmp_ref_inc(route_info.nhop_h);
-    }
-  }
-}
-
-/*
- * Routine Description:
- *    Delete route and remove entry from the database
- *
- * Arguments:
- *    [in] vrf_h - vrf handle
- *    [in] dst - IP address associated with route
- *
- * Return Values:
- *    void
- */
-
-void route_delete(switchlink_handle_t vrf_h, switchlink_ip_addr_t *dst) {
-  bool ecmp_enable = false;
-  switchlink_handle_t ecmp_h;
-
-  if (!dst) {
-    return;
-  }
-
-  switchlink_db_status_t status;
-  switchlink_db_route_info_t route_info;
-  memset(&route_info, 0, sizeof(switchlink_db_route_info_t));
-  route_info.vrf_h = vrf_h;
-  memcpy(&(route_info.ip_addr), dst, sizeof(switchlink_ip_addr_t));
-  status = switchlink_db_route_get_info(&route_info);
-  if (status != SWITCHLINK_DB_STATUS_SUCCESS) {
-    return;
-  }
-
-  dzlog_info("Delete route: 0x%x/%d", dst->ip.v4addr.s_addr,
-             dst->prefix_len);
-  if (switchlink_route_delete(&route_info) == -1) {
-    return;
-  }
-
-  ecmp_enable = route_info.ecmp;
-  ecmp_h = route_info.nhop_h;
-
-  status = switchlink_db_route_delete(&route_info);
-  if (status == SWITCHLINK_DB_STATUS_SUCCESS) {
-    if (ecmp_enable) {
-      ecmp_delete(ecmp_h);
-    }
-  }
-}
 
 /*
  * Routine Description:
@@ -292,7 +77,7 @@ static switchlink_handle_t process_ecmp(uint8_t family,
           nexthop_info.using_by |= SWITCHLINK_NHOP_FROM_ROUTE;
           switchlink_db_nexthop_update_using_by(&nexthop_info);
         } else {
-          if (!switchlink_nexthop_create(&nexthop_info)) {
+          if (!switchlink_create_nexthop(&nexthop_info)) {
              dzlog_debug("Created nhop 0x%lx handler, update from"
                       " route", nexthop_info.nhop_h);
              ecmp_info.nhops[ecmp_info.num_nhops] = nexthop_info.nhop_h;
@@ -315,7 +100,7 @@ static switchlink_handle_t process_ecmp(uint8_t family,
 
   status = switchlink_db_ecmp_get_info(&ecmp_info);
   if (status == SWITCHLINK_DB_STATUS_ITEM_NOT_FOUND) {
-    switchlink_ecmp_create(&ecmp_info);
+    switchlink_create_ecmp(&ecmp_info);
     switchlink_db_ecmp_add(&ecmp_info);
   }
 
@@ -477,14 +262,14 @@ void process_route_msg(struct nlmsghdr *nlmsg, int type) {
                                                      dst_valid ?
                                                      dst_addr.ip.v4addr.s_addr :
                                                      0);
-    route_create(g_default_vrf_h,
-                 (dst_valid ? &dst_addr : NULL),
-                 (gateway_valid ? &gateway_addr : NULL),
-                 ecmp_h,
-                 ifinfo.intf_h);
+    switchlink_create_route(g_default_vrf_h,
+                            (dst_valid ? &dst_addr : NULL),
+                            (gateway_valid ? &gateway_addr : NULL),
+                            ecmp_h,
+                            ifinfo.intf_h);
   } else {
     dzlog_info("Delete route with addr: 0x%x", dst_valid ?
                                              dst_addr.ip.v4addr.s_addr : 0);
-    route_delete(g_default_vrf_h, (dst_valid ? &dst_addr : NULL));
+    switchlink_delete_route(g_default_vrf_h, (dst_valid ? &dst_addr : NULL));
   }
 }
