+++ switchlink/switchlink_utils.h
@@ -15,17 +15,17 @@
  * limitations under the License.
  */
 
-#ifndef __SWITCHLINK_INT_H__
-#define __SWITCHLINK_INT_H__
+#ifndef __SWITCHLINK_UTILS_H__
+#define __SWITCHLINK_UTILS_H__
 
-struct nlmsghdr;
-extern void switchlink_init_db(void);
-extern void switchlink_init_api(void);
-extern void switchlink_init_link(void);
 
-extern void process_link_msg(struct nlmsghdr *nlmsg, int type);
-extern void process_neigh_msg(struct nlmsghdr *nlmsg, int type);
-extern void process_address_msg(struct nlmsghdr *nlmsg, int type);
-extern void process_route_msg(struct nlmsghdr *nlmsg, int type);
+#include <stdint.h>
+#include <stdbool.h>
+#include <netlink/msg.h>
+#include <netlink/netlink.h>
 
-#endif /* __SWITCHLINK_INT_H__ */
+#include "switchutils/switch_utils.h"
+
+uint32_t ipv4_prefix_len_to_mask(uint32_t prefix_len);
+struct in6_addr ipv6_prefix_len_to_mask(uint32_t prefix_len);
+#endif /* __SWITCHLINK_UTILS_H__ */
